setup client
