+++ client/.babelrc
@@ -1,4 +1,7 @@
 {
   "presets": ["@babel/preset-env", "@babel/preset-react"],
-  "plugins": ["transform-class-properties"]
+  "plugins": [
+    "transform-class-properties",
+    "@babel/plugin-proposal-object-rest-spread"
+  ]
 }
