+++ client/jest.config.json
@@ -1,0 +1,4 @@
+{
+  "setupFiles": ["raf/polyfill", "<rootDir>/src/tests/setupTests.js"],
+  "snapshotSerializers": ["enzyme-to-json/serializer"]
+}
