+++ client/package.json
@@ -7,13 +7,26 @@
   "license": "MIT",
   "scripts": {
     "build": "webpack ",
-    "dev-server": "webpack-dev-server"
+    "dev-server": "webpack-dev-server",
+    "test": "jest --config=jest.config.json --watch"
   },
   "dependencies": {
     "babel-core": "^6.26.3",
     "babel-loader": "^8.0.6",
+    "css-loader": "^3.2.0",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.15.1",
+    "enzyme-to-json": "^3.4.3",
+    "jest": "^24.9.0",
+    "node-sass": "^4.13.0",
+    "raf": "^3.4.1",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
+    "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
+    "redux": "^4.0.4",
+    "sass-loader": "^8.0.0",
+    "style-loader": "^1.0.0",
     "webpack-dev-server": "^3.9.0"
   },
   "devDependencies": {
