+++ client/src/components/Profile.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Profile = props => {
+  return (
+    <>
+      <h1>This is a profile</h1>
+    </>
+  );
+};
+
+export default Profile;
