+++ client/src/routers/AppRouter.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
+
+import Dashboard from '../components/Dashboard';
+
+const AppRouter = () => (
+  <BrowserRouter>
+    <Switch>
+      <Route path="/" component={Dashboard} />
+    </Switch>
+  </BrowserRouter>
+);
+
+export default AppRouter;
