+++ client/src/styles/base/_base.scss
@@ -1,0 +1,7 @@
+html {
+  font-size: 62.5%;
+}
+body {
+  font-family: Helvetica, Arial, sans-serif;
+  font-size: 1.6rem;
+}
