+++ client/src/styles/base/_settings.scss
@@ -1,0 +1,6 @@
+//colors
+$theme-color: blue;
+
+//m-size
+
+$m-size: 1.6rem;
