+++ client/src/styles/main.scss
@@ -1,0 +1,7 @@
+@import './base/settings';
+@import './base/base';
+@import './components/navbar';
+
+* {
+  color: $theme-color;
+}
